fix(core): move MetricPoint to the port it is spoken in
The import contract caught the hexagon inverted: domain.contracts.metrics_history
imported MetricPoint from infrastructure, because that is where the record was
originally defined next to its SQLite store.

It is the record the port speaks in, so both backends and every caller need it,
and a port reaching into one adapter for its own vocabulary is the wrong
direction regardless of which adapter got there first. Moved to the port; the
SQLite module re-exports the name so existing imports keep working.
