+++ src/mcp_hangar/domain/contracts/metrics_history.py
@@ -1,19 +1,36 @@
 """Where per-server metric snapshots are kept for the history endpoints.
 
 A port for the same reason the saga store has one: it is state the gateway
-persists, so the selected storage backend has to provide it. `MetricPoint`
-stays where it is defined and is imported here rather than redeclared -- a
-second definition of the same record is how two halves of a system stop
-agreeing about what a metric is.
+persists, so the selected storage backend has to provide it.
+
+`MetricPoint` lives here rather than in the SQLite adapter it was born in. It is
+the record the port speaks in, so both backends and every caller need it, and a
+port that reaches into an adapter for its own vocabulary is the hexagon
+inverted -- the import contract said so out loud. The SQLite module re-exports
+the name, so existing imports keep working.
 """
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING
+from dataclasses import dataclass
 
-if TYPE_CHECKING:
-    from mcp_hangar.infrastructure.persistence.metrics_history_store import MetricPoint
+
+@dataclass
+class MetricPoint:
+    """A single time-series data point.
+
+    Attributes:
+        mcp_server_id: McpServer this metric belongs to.
+        metric_name: Name of the metric (e.g. ``tool_calls_total``).
+        value: Numeric value.
+        recorded_at: Unix timestamp (seconds since epoch) when the snapshot was taken.
+    """
+
+    mcp_server_id: str
+    metric_name: str
+    value: float
+    recorded_at: float
 
 
 class IMetricsHistoryStore(ABC):
